Reduce large copies of HD dat file structures
Previously, we'd read the `aocDat` and `hdDat` structures, then do a lot
of assignments to `aocDat`. This required:
- Destroying all of the structures in `aocDat`
- Copying all the data from `hdDat` (copy constructing)
- Destroying all of the structures in `hdDat`

Which would add up to a significant amount of work.

With std::move, we now only:
- Destroy all of the structures in `aocDat`
- Change the pointers in `aocDat` to point to `hdDat` and clear the
pointers in `hdDat` (super cheap)

Eventually `hdDat` is also destroyed but it only requires destructing
the small amount of data it still holds.

In my unscientific test, with the old method `transferHdDatElements`
took about 5% of install time. Now, it takes about 1%.